requirement.txt
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SocketIO==3.1.2
opencv-python==4.1.1.26
lxml
tqdm
tensorflow==2.3.0rc0
absl-py
easydict
matplotlib
pillow
pandas
numpy
requests
scikit-image==0.14.2
scikit-learn==0.21.3
pyarrow
datetime
boto3
pickle-mixin
gevent
gunicorn
xlrd
